correcting joinalble data structure
The data structure was wrong because we could not have the same line
with the same key in a joinable. This is why the map changed to a list
of two elements - key and a record.
I think that the word key is wrong and should be changed to something
else.